\name{macro.data.counts}
\alias{macro.data.counts}
\title{Macrophage data set.}
\description{Showcase example: Macrophage responses to stimulation.}
\usage{
data(macro.data.counts)
}
\format{
The data set is stored in RData (binary) format. It provides transcripts' read counts as numeric matrix.
}
\details{
The data set was downloaded from Gene Expression Omnibus repository
(http://www.ncbi.nlm.nih.gov/geo, GEO accession no. GSE79363).
In 60 cells about 44,000 transcripts were measured using RNA-seq (Illumina MiSeq). 
}
\source{
http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE79363
}