% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRAT.R
\docType{package}
\name{SCRAT}
\alias{SCRAT}
\alias{SCRAT-package}
\title{SCRAT: Single-Cell Regulome Analysis Tool}
\description{
SCRAT is a comprehensive analysis tool for single-cell regulome data.
}
\details{
SCRAT provides essential tools for users to read in single-cell regolome data (ChIP-seq, ATAC-seq, DNase-seq) and summarize into different types of features. It also allows users to visualize the features, cluster samples and identify key features.
}

