% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRATsummary.R
\name{SCRATsummary}
\alias{SCRATsummary}
\title{SCRATsummary}
\usage{
SCRATsummary(dir, genome, bamfile = NULL, singlepair = "automated",
  removeblacklist = T, log2transform = T, featurelist = c("GENE", "ENCL",
  "MOTIF_TRANSFAC", "MOTIF_JASPAR", "GSEA"), Genestarttype = "TSSup",
  Geneendtype = "TSSdown", Genestartbp = 3000, Geneendbp = 1000,
  ENCLclunum = 2000, Motifflank = 100, GSEAterm = "c5.bp",
  GSEAstarttype = "TSSup", GSEAendtype = "TSSdown", GSEAstartbp = 3000,
  GSEAendbp = 1000)
}
\arguments{
\item{dir}{The folder where the bam files are stored. All bam files within the folder will be analyzed. Only useful when bamfile is NULL.}

\item{genome}{The mapped genome of the bam files. Should be one the following: "hg19", "hg38", "mm9", "mm10"}

\item{bamfile}{A character vector of the path to bam files. If not NULL, dir will not be used.}

\item{singlepair}{Whether the original sequencing files are single-end or paired-end. Should be one of the following: "automated", "single", "pair". Default is "automated" where SCRAT will automatically determine the type.}

\item{removeblacklist}{Logical value indicating whether black list regions should be removed.}

\item{log2transform}{Logical value indicating whether the read counts should be log2 transformed (after adding pseudo-count of 1).}

\item{featurelist}{A character vector specifying what kind of features should be considered. Should be from the following: "GENE","ENCL","MOTIF_TRANSFAC","MOTIF_JASPAR","GSEA". By default all features are included. Note that "GSEA" features could be slow to run.}

\item{Genestarttype}{For "GENE" features, type of starting site. Should be one of the following: "TSSup", "TSSdown", "TESup", "TESdown". The four options stands for TSS upstream, TSS downstream, TES upstream and TES downstream.}

\item{Geneendtype}{For "GENE" features, type of ending site. Options same as Genestarttype}

\item{Genestartbp}{For "GENE" features, how many base pairs away from starting TSS/TES. For example, Genestarttype="TSSup" and Genestartbp=500 means 500 bp upstream of TSS.}

\item{Geneendbp}{For "GENE" features, how many base pairs away from ending TSS/TES.}

\item{ENCLclunum}{Number of clusters for ENCL features. Should be one of 1000, 2000 and 5000}

\item{Motifflank}{Defines the size of flanking region of motif sites in base pairs.}

\item{GSEAterm}{The GSEA terms included in the analysis. Only useful when "GSEA" is included in featurelist. Should be one of the following: "h.all","c1.all","c2.cgp","c2.cp","c3.mir","c3.tft","c4.cgn","c4.cm","c5.bp","c5.cc","c5.mf","c6.all","c7.all".}

\item{GSEAstarttype}{For "GSEA" features, type of starting site.}

\item{GSEAendtype}{For "GSEA" features, type of ending site.}

\item{GSEAstartbp}{For "GSEA" features, how many base pairs away from starting TSS/TES.}

\item{GSEAendbp}{For "GSEA" features, how many base pairs away from ending TSS/TES.}
}
\description{
Compile SCRAT summary table
}
\details{
This function will compile a SCRAT summary table from bam files. The results should be the same as run on GUI.
}
\examples{
\dontrun{
   SCRATsummary(dir="bamfiledir",genome="hg19")
}
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
