% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRATui.R
\name{SCRATui}
\alias{SCRATui}
\title{SCRATui}
\usage{
SCRATui()
}
\description{
Launch the SCRAT user interface in local machine
}
\details{
This function will automatically launch the SCRAT user interface in a web browser. 
The user interface can also be accessed by http://zhiji.shinyapps.io/SCRAT. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally for faster running speed.
}
\examples{
\dontrun{
   SCRATui()
}
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
